Add custom json canonicalization function
The in-toto reference implementation uses a custom json
canonicalization function from securesystemslib.

Unfortunately it is not compatible with a similar go implementation
See https://github.com/secure-systems-lab/securesystemslib/issues/159

In order to verify signatures created by the reference
implementation and also to create keyids, we have to use the
same canonicalization, hence this commit ports securesystemslib's
encode_canonical to go.