Add basic test for InTotoVerify (WIP)
Adds a basic test for verifylib that calls InTotoVerify, passing
the path to a signed layout and loads and passes the corresponding
public key.

InTotoVerify then loads the layout, verifies the signature and
loads the corresponding link files.

NOTE:
I'd rather have this in ROOT/test, but I have to understand better
how testing (and importing the modules) works first.