Adapt metablock load error handling in InTotoVerify
Like in the reference implementation the library functions should
not handle/recover if the call needs to exit but should rather
pass through the error to the caller.

The caller, e.g. a command line tool should do the final error
handling.

The commit also adds a unittest for InTotoVerify with where the
layout can't be loaded.