Add basic runlib
Runlib provides an InTotoRun function to execute the passed command
(in a subproces) creating unsigned link metadata recording
materials, products and byproducts.

InTotoRun cannot (yet) be used to provide signed evidence for
a step of a layout. The current runlib implementation may only be
used to create link metadata for inspections of a layout at the
time of layout verification.