Use newly added Set util in VerifyArtifacts
Use the Set struct and its set function instead of a string slice
for the artifacts queue and artifact consumption by rules.

This commit also updates the VerifyArtifacts helper verifyMatchRule
to return consumed artifacts instead of the updated artifacts
queue, to align better with the corresponding pseudo code of the
specification.