Fix error handling in _encodeCanonical
_encodeCanonical is a recursive function that uses panic to error
out and defer to recover from the error and set the returned error.

Prior to this change, the intended behavior only worked on the
first recursion level, because the errors were not passed back up
the recursion stack, when returned from a recursive call.

This commit fixes this, by immediately returning an error if
returned from a recursive call.