Prohibit floating point numbers in EncodeCanonical
Prior to this change any numeric values were converted to
integers during canonicalization. This does not conform with the
OLPC canonical JSON specification, which disallows floating point
numbers.

This commit changes `EncodeCanonical` to decode any numeric values
as `json.Number` (instead of the default `float64`), and
`_encodeCanonical` to fail if any `json.Number` is not a whole
number.

Note that the json decoder in `Metablock.load` in `models` is not
changed, and will decode any numeric values from a loaded json file
as `float64`. However, these numbers may still be successfully
encoded with `EncodeCanonical` as long as they hold whole numbers
(see tests).