Fix inconsistent link dumping/loading
Before this commit we used []byte64 and int in our in-memory
link representation. This lead to numerous issues:

1. Using []byte64 for stderr/stdout meant, that we dump them as base64
in our JSON file. This was inconsistent to our in-toto python
implementation, that stores output as strings in JSON files.

2. Go unmarshalls a number as float64, therefore we can't easily
store the return-value as integer, although an integer would be a better
choice. Storing it as a integer, would cost rewrites of the complete
model and model testing.