add ecdsa support
This commit adds support for the ecdsa key scheme.
We should support all FIPS 186-3 curves out of the box.
However, we must note, that if we ever upgrade our hashing
algorithm from SHA256, the code will get more complex, because
the hash size must satisfy the curve size, otherwise the
hash may get truncated. The latter could result in a security vulnerability (forming a hash, which truncated part is equal to the truncated part of the to verified bytes). Furthermore,
we do no curve detection here and just save the signature parts r and s into a byte slice without a fixed length. Also it's still unclear if r and s are always the same of the size