implement proper ecdsa signature encoding
We now use ASN1.DER for encoding the ecdsa signature parameters
r and s. This fixes our problems with different r and s lengths.
Furthermore it is the same way to deal with the signature
as in our securesystemslib and it is therefore the interoperable
way to handle an ecdsa signature. With Go 1.15 we might can switch
to new ecdsa ASN1Sign methods, if necessary.