set correct ecdsa scheme
We set the correct ecdsa schemes as constants. Right now we don't
really differ between curves, because Go's crypto/(ecdsa|x509)
are fully transparent with curves. If we want to differ between
curves we might need to add addition logic, right now it's up
to the developer, to choose the right scheme for the right curve.