Call pathspec.GitIgnore directly and add test cases
This commit removes our applyGitIgnorePatterns function, because
the position of it has been wrong. We need to invoke pathspec
for every path in our filewalk. With the new position the dedicated
function is no longer needed and we can directly call pathspec.GitIgnore.

Furthermore we add a function for creating more complex test cases and
a copy helper function. This will allow us to add more complex test
cases in the future (if needed).