Do not hard-code test layout shasum in tests
The test layout shasum is prone to change, e.g. when the layout's
expiration time needs to be updated and the document re-singed (see
previous commit).  Hence it is not practical to hard-code the
layout's shasum, as was done prior to this commit in order to test
in-toto artifact recording.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>