Rename test demo.layout.template to demo.layout
The demo layout in the Python reference implementation's test data
uses the ".template" suffix to indicate that it lacks the
expiration field and signatures, which are added on the fly during
tests.

The test layout used here is fully in-toto conformant, as a
consequence the .template suffix is not needed. This is especially
useful for the (tested) example snippet, which has an educational
purpose (see examples_test.go).

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>