Adjust how certificates are loaded
The hacky code I wrote to load certificates caught up with us.  When
verifying a layout whose public key is a certificate we wouldn't
properly generate an id for the key, which meant in-toto verify could
never match a signature with the provided certificate.  I've reworked
how certificates are loaded a bit to act in a more expected matter.

We still have to ad-hoc attach the certificate's bytes when executing
in-toto run due to the way setKeyComponents work.  I'd like to make this
a smoother process but for the time being this is less hacky overall
than how it was being done before.

Use example.layout.cert.pem in test-verify target.

test-sign signs with example.layout.cert.pem but test-verify was using
root.cert.pem to verify

Add build target to test-sign and test-verify

test-verify and test-sign error out on clean repo due to build not being
run until test-run

Fix some typos and left over debug statements

clean up Makefile