Add slsa0.1 statement struct back
As per https://github.com/in-toto/in-toto-golang/issues/176, we want to
maintain multiple versions of slsa-specific statement struct especially
when slsa v1 comes out in next couple of months. Reverting slsa0.1 back
helps the community evaluate the work.

Summary:
- Version specific statement structs are added `ProvenanceStatementSLSA01`
and `ProvenanceStatementSLSA02`.
- The old `ProvenanceStatement` struct that is specific to slsa 0.2 is
still available, but will be officially deprecated in favour of
version explicit alternatives.
- Some common types shared by 0.1 and 0.2 are moved to the newly added
`common` package including `DigestSet`, `ProvenanceBuilder` and `ProvenanceMaterial`.
This might be a breaking change for the downstream users who use those
types to construct a statement.

Signed-off-by: Chuang Wang <chuangw@google.com>